requirements.txt
dlib==19.17.0
dpkt==1.9.3
future==0.18.2
imutils==0.5.3
iso8601==0.1.12
numpy==1.17.1
opencv-contrib-python==3.4.4.19
pbr==5.4.2
pypcap==1.2.3
pyserial==3.4
PyYAML==5.1.2
serial==0.0.97
simple-pid==0.2.3
six==1.12.0
stevedore==1.30.1
virtualenv==16.7.4
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
